+++ .github/workflows/ci.yml
@@ -49,6 +49,8 @@ jobs:
           - test --profile=ci-dev -p cairo-lang-sierra-to-casm
           - test --profile=ci-dev -p cairo-lang-sierra-type-size
           - test --profile=ci-dev -p cairo-lang-starknet
+          - test --profile=ci-dev -p cairo-lang-starknet-types --features=serde 
+          - test --profile=ci-dev -p cairo-lang-starknet-types --no-default-features --features=serde 
           - test --profile=ci-dev -p cairo-lang-syntax
           - test --profile=ci-dev -p cairo-lang-syntax-codegen
           - test --profile=ci-dev -p cairo-lang-test-runner
