+++ crates/bin/starknet-sierra-compile/src/main.rs
@@ -2,7 +2,7 @@ use std::fs;
 
 use anyhow::Context;
 use cairo_lang_starknet::allowed_libfuncs::{validate_compatible_sierra_version, ListSelector};
-use cairo_lang_starknet::casm_contract_class::CasmContractClass;
+use cairo_lang_starknet::casm_contract_class_from_contract_class::casm_contract_class_from_contract_class;
 use cairo_lang_starknet::contract_class::{ContractClass, ContractEntryPoints};
 use cairo_lang_utils::bigint::BigUintAsHex;
 use clap::Parser;
@@ -64,7 +64,7 @@ fn main() -> anyhow::Result<()> {
     };
     validate_compatible_sierra_version(&contract_class, list_selector)?;
     let casm_contract =
-        CasmContractClass::from_contract_class(contract_class, args.add_pythonic_hints)
+        casm_contract_class_from_contract_class(contract_class, args.add_pythonic_hints)
             .with_context(|| "Compilation failed.")?;
 
     let res = serde_json::to_string_pretty(&casm_contract)
