+++ crates/cairo-lang-starknet-types/src/lib.rs
@@ -1,0 +1,9 @@
+#![cfg_attr(not(feature = "std"), no_std)]
+
+#[cfg(not(feature = "std"))]
+extern crate alloc;
+
+pub mod abi;
+pub mod casm_contract_class;
+pub mod compiler_version;
+pub mod contract_class;
