+++ crates/cairo-lang-starknet/src/allowed_libfuncs.rs
@@ -95,8 +95,8 @@ mod serde_utils {
     use std::fs;
 
     use super::*;
-    use crate::contract_class::ContractClass;
     use crate::felt252_serde::sierra_from_felt252s;
+    use cairo_lang_starknet_types::contract_class::ContractClass;
 
     pub fn lookup_allowed_libfuncs_list(
         list_selector: ListSelector,
