+++ crates/cairo-lang-starknet/src/contract.rs
@@ -34,7 +34,7 @@ use sha3::{Digest, Keccak256};
 use {cairo_lang_lowering as lowering, cairo_lang_semantic as semantic};
 
 use crate::aliased::Aliased;
-use crate::contract_class::{extract_semantic_entrypoints, SemanticEntryPoints};
+use crate::compile_contract_class::{extract_semantic_entrypoints, SemanticEntryPoints};
 use crate::plugin::aux_data::StarkNetContractAuxData;
 use crate::plugin::consts::{ABI_ATTR, ABI_ATTR_EMBED_V0_ARG};
