+++ crates/cairo-lang-starknet/src/plugin/events.rs
@@ -11,7 +11,7 @@ use smol_str::SmolStr;
 
 use super::consts::{EVENT_ATTR, EVENT_TRAIT, EVENT_TYPE_NAME};
 use super::starknet_module::StarknetModuleKind;
-use crate::abi::EventFieldKind;
+use cairo_lang_starknet_types::abi::EventFieldKind;
 
 /// Generated auxiliary data for the `#[derive(starknet::Event)]` attribute.
 #[derive(Clone, Debug, PartialEq, Eq)]
